+++ client/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## Unreleased
+
+### Added
+
+### Fixed
+
+- transport selection order
+- transports not expanden in view
+- transport tasks
+
 ## 0.5.22 2019-03-26
 
 ### Added